change the receipt of the order price including VAT and delivery

+++ next-pizza-api/src/main/java/ru/alex/nextpizzaapi/database/entity/Order.java
@@ -28,7 +28,7 @@ public class Order extends AuditableEntity {
     private String token;
 
     @Column(name = "total_amount")
-    private Integer totalAmount;
+    private Float totalAmount;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status")
